+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+* bump `esp-hal` to `1.0.0-rc.1` [#34](https://github.com/liebman/esp-hub75/pull/34)
+* bump `hub75-framebuffer` to `0.5.0` [#34](https://github.com/liebman/esp-hub75/pull/34)
+
 ## [0.6.0] - 2025-08-17
 
 ### Added